whatsapp_client/requirements.txt
fastapi 
uvicorn 
twilio 
redis 
python-multipart 
python-dotenv
httpx
ngrok